fix(lint): clean up pre-existing unconvert/unused findings in linux tests
hdr.DataEnd is already uint32, so the wrapping uint32() conversion in
e2e_linux_test.go was a no-op (unconvert), and childDefined in
child_reader_linux_test.go was declared but never read (unused). These
predate this session's changes (introduced with the Linux e2e tests) and
have been silently failing the Lint job on every push since, just masked
by the Coverage gate job failing earlier in the pipeline.

+++ pkg/featcache/e2e_linux_test.go
@@ -142,7 +142,7 @@ func TestE2ESegmentLayoutConsistency(t *testing.T) {
 		t.Fatalf("hash table overlaps data region")
 	}
 	// First data bytes are initialized (zeroed hash table, empty data).
-	if uint32(hdr.DataEnd) != hdr.DataOffset {
+	if hdr.DataEnd != hdr.DataOffset {
 		t.Fatalf("DataEnd = %d, want %d", hdr.DataEnd, hdr.DataOffset)
 	}
 }
